bart/requirements.txt
PyTorch version >= 1.4.0
Python version >= 3.6
